feat: add useParsedValue hook for parsing variable strings
